refact ; notification class cleanup
