+++ src/Kanvas/Notifications/DataTransferObject/Notifications.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Kanvas\Notifications\DataTransferObject;
 
-class Notifications
+use Spatie\LaravelData\Data;
+
+class Notifications extends Data
 {
     /**
      * __construct.
@@ -21,36 +23,7 @@ class Notifications
         public int $entity_id,
         public string $content,
         public int $read,
-        public string $created_at,
-        public ?string $updated_at = null,
-        public int $is_deleted = 0,
         public ?string $content_group = null,
     ) {
-    }
-
-    /**
-     * fromArray.
-     *
-     * @param  array $data
-     *
-     * @return self
-     */
-    public static function fromArray(array $data): self
-    {
-        return new self(
-            $data['users_id'],
-            $data['from_users_id'],
-            $data['companies_id'],
-            $data['apps_id'],
-            $data['system_modules_id'],
-            $data['notification_type_id'],
-            $data['entity_id'],
-            $data['content'],
-            $data['read'],
-            $data['created_at'],
-            $data['updated_at'] ?? null,
-            $data['is_deleted'] ?? 0,
-            $data['content_group'] ?? null,
-        );
     }
 }
