+++ src/Kanvas/Notifications/Notification.php
@@ -7,7 +7,6 @@ namespace Kanvas\Notifications;
 use Baka\Contracts\AppInterface;
 use Baka\Support\Str;
 use Baka\Users\Contracts\UserInterface;
-use Exception;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Database\Eloquent\Model;
@@ -17,21 +16,23 @@ use Kanvas\Apps\Models\Apps;
 use Kanvas\Notifications\Channels\KanvasDatabase as KanvasDatabaseChannel;
 use Kanvas\Notifications\Interfaces\EmailInterfaces;
 use Kanvas\Notifications\Models\NotificationTypes;
+use Kanvas\Notifications\Traits\NotificationRenderTrait;
+use Kanvas\Notifications\Traits\NotificationStorageTrait;
 use Kanvas\SystemModules\Repositories\SystemModulesRepository;
-use Kanvas\Templates\Actions\RenderTemplateAction;
 use Kanvas\Users\Models\Users;
 
 class Notification extends LaravelNotification implements EmailInterfaces, ShouldQueue
 {
     use Queueable;
+    use NotificationStorageTrait;
+    use NotificationRenderTrait;
 
     protected Model $entity;
     protected AppInterface $app;
     protected ?NotificationTypes $type = null;
-    protected ?string $templateName = null;
     protected ?UserInterface $fromUser = null;
     protected ?UserInterface $toUser = null;
-    public array $data = [];
+
     public array $via = [
         KanvasDatabaseChannel::class,
     ];
@@ -84,103 +85,6 @@ class Notification extends LaravelNotification implements EmailInterfaces, Shoul
         return (new MailMessage())
                 ->from($fromEmail, $fromName)
                 ->view('emails.layout', ['html' => $this->message()]);
-    }
-
-    /**
-     * toKanvasDatabase.
-     */
-    public function toKanvasDatabase(UserInterface $notifiable): array
-    {
-        $this->toUser = $notifiable;
-
-        try {
-            $fromUserId = $this->getFromUser()->getId();
-        } catch (Exception $e) {
-            //for now, we need to clean this up -_-
-            $fromUserId = 0;
-        }
-
-        return [
-            'users_id' => $notifiable->getId(),
-            'from_users_id' => $fromUserId,
-            'companies_id' => $notifiable->getCurrentCompany()->getId(),
-            'apps_id' => $this->app->getId(),
-            'system_modules_id' => $this->getType()->system_modules_id,
-            'notification_type_id' => $this->getType()->getId(),
-            'entity_id' => method_exists($this->entity, 'getId') ? $this->entity->getId() : $this->entity->id,
-            'content' => $this->message(),
-            'read' => 0,
-            'created_at' => date('Y-m-d H:i:s'),
-            'updated_at' => date('Y-m-d H:i:s'),
-            'is_deleted' => 0,
-        ];
-    }
-
-    /**
-     * Notification message the user will get.
-     */
-    public function message(): string
-    {
-        if ($this->getType()->hasEmailTemplate()) {
-            return $this->getEmailTemplate();
-        }
-
-        return '';
-    }
-
-    /**
-     * Given the HTML for the current email notification
-     */
-    protected function getEmailTemplate(): string
-    {
-        if (! $this->getType()->hasEmailTemplate()) {
-            throw new Exception('This notification type does not have an email template');
-        }
-
-        $renderTemplate = new RenderTemplateAction($this->app);
-
-        return $renderTemplate->execute(
-            $this->getTemplateName(),
-            $this->getData()
-        );
-    }
-
-    /**
-     * setTemplateName
-     *
-     * @param  mixed $name
-     */
-    public function setTemplateName(string $name): self
-    {
-        $this->templateName = $name;
-
-        return $this;
-    }
-
-    /**
-     * setData
-     */
-    public function setData(array $data): self
-    {
-        $this->data = $data;
-
-        return $this;
-    }
-
-    /**
-     * getData.
-     */
-    public function getData(): array
-    {
-        return $this->data;
-    }
-
-    /*
-    * Get notification template Name
-    */
-    public function getTemplateName(): ?string
-    {
-        return $this->templateName === null ? $this->getType()->template : $this->templateName;
     }
 
     /**
